\name{details}
\alias{details}
\docType{data}
\title{
Template for use in function allmetap
}
\description{
Provides a template for the output of allmetap}
\usage{details}
\format{
  A data frame with 8 observations on the following 2 variables.
  \describe{
    \item{\code{funcs}}{a character vector}
    \item{\code{eponyms}}{a character vector}
  }
}
\keyword{internal}
