\name{metap-internal}
\alias{metap-internal}
\alias{docontours}
\alias{makexscale}
\alias{makeyscale}
\title{Internal functions}
\description{Functions for internal use}
\usage{
docontours(conttype, xvals, contvals, xtrans, ytrans,
   xlims, ylims, ltys, contlabs)
makeyscale(yscale, userfun)
makexscale(p)
}
\details{Not to be called by the user}
\keyword{internal}
