\name{metap-package}
\alias{metap-package}
\alias{metap}
\docType{package}
\title{\packageTitle{metap}}
\description{\packageDescription{metap}}
\details{
\packageIndices{metap}

Provides a number of ways in which significance levels may be combined
in a meta-analysis and includes most ot the methods
in \insertCite{becker94;textual}{metap}.
It includes a number of datasets taken from the literature. 
It also provides a display and
an informal graphical test due to Schweder
and Spjotvoll \insertCite{schweder82}{metap}
and the lowest slope line of Benjamini and
Hochberg \insertCite{benjamini00}{metap}.
The albatros plot of Harrison et al \insertCite{harrison17}{metap}
is also provided.

}
%\author{\packageAuthor{metap}}

%\packageMaintainer{metap}

\references{
\insertAllCited{}
}
\keyword{ package }
\seealso{
The issue of meta-analysis of signficance levels is not completely
unconnected with the topic of adjustment for multiple
comparisons as in for example \code{\link{p.adjust}}
}
